+++ vsax/ops/__init__.py
@@ -3,11 +3,13 @@
 from vsax.ops.binary import BinaryOperations
 from vsax.ops.fhrr import FHRROperations
 from vsax.ops.map import MAPOperations
-from vsax.ops.quaternion import QuaternionOperations
+from vsax.ops.quaternion import QuaternionOperations, sandwich, sandwich_unit
 
 __all__ = [
     "FHRROperations",
     "MAPOperations",
     "BinaryOperations",
     "QuaternionOperations",
+    "sandwich",
+    "sandwich_unit",
 ]